+++ packages/axway-cli-pm/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v4.0.2 (Nov 5, 2021)
+
+ * chore: Updated dependencies.
+
 # v4.0.1 (Oct 18, 2021)
 
  * chore: Updated dependencies.